requirements.txt
streamlit==1.28.0
pdfplumber==0.10.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0